fix: format function calls for improved readability in layout generation

+++ src/scenes/stage/systems/mod.rs
@@ -205,7 +205,13 @@ fn populate_stage_contents(
                 stage_root,
                 tiled_map_assets,
                 viewport,
-                tile_position_to_world((x as f32, y as f32), real_tile_size, viewport_size, scale, 2.0),
+                tile_position_to_world(
+                    (x as f32, y as f32),
+                    real_tile_size,
+                    viewport_size,
+                    scale,
+                    2.0,
+                ),
             );
         });
 }
@@ -218,8 +224,8 @@ fn tile_position_to_world(
     adjust_y: f32,
 ) -> (f32, f32, f32) {
     (
-        (tile_pos.0 as f32 + 0.5) * tile_size.x - viewport_size.x / 2.0,
-        (tile_pos.1 as f32 + 0.5) * tile_size.y - adjust_y - viewport_size.y / 2.0,
+        (tile_pos.0 + 0.5) * tile_size.x - viewport_size.x / 2.0,
+        (tile_pos.1 + 0.5) * tile_size.y - adjust_y - viewport_size.y / 2.0,
         scale,
     )
 }
